+++ src/package/vendors/pptx/worker/index.ts
@@ -1,8 +1,9 @@
 import { refWorker } from '@/package/common/worker-ref'
+import PptxWorker from './pptx.worker.js?worker&inline'
 
 export default {
   create() {
     return refWorker('pptx.worker.js').defaults(() =>
-      new Worker(new URL('./pptx.worker.js', import.meta.url), { type: 'module' }))
+      new PptxWorker())
   }
 }